subsys: settings: nvs: use dedicated lookup cache hash function
NVS lookup cache currently uses CRC8/16 as a hash function
to determine the cache position. However, it may result in
costly collisions between rarely written keys and more
frequently written values. Introduce a dedicated hash
function to be used when NVS is used as the settings backend.

Also, add a test case to verify that the default hash
function is not extremely bad.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>